+++ eras/alonzo/test-suite/test/Test/Cardano/Ledger/Alonzo/Golden.hs
@@ -298,7 +298,7 @@ goldenMinFee =
           pricesParam = Prices priceMem priceSteps
           pp =
             emptyPParams
-              & ppMinFeeAL .~ Coin 44
+              & ppMinFeeAL .~ CoinPerByte (Coin 44)
               & ppMinFeeBL .~ Coin 155381
               & ppPricesL .~ pricesParam